BraveSessionCache: Store random state in the callback.
Replace the local static variable for random state in
AudioFarblingCallback with a copy of the pseudo-random
generator object bound to the callback itself.

This is less likely to cause problems later if multiple
callbacks end up with interleaved invocations.

As far as I can tell, the bound argument needs to be a
shared_ptr because there's no other way for the wrapped
function to obtain mutable access to bound copy, which
is essential here to advance the sequence.