WIP: flip a pref from the settings page.
Using a common pref seems reasonable, although it's a little
strange they aren't namespaced.

There's a `brave/components/variations` subdir, but it only
contains a build flag, no code, so I just added the the
pref registration to the main process function.

We've always discussed this as an opt-out. so register the
pref with default value of `true`.

BTW, the p3a pref has an extra change function; I'm not sure if
we need that. Or if our pref needs to be global instead of local
to the profile.