WIP: Register the variations pref
Using a common pref seems reasonable, although it's a little
strange they aren't namespaced.

There's a `brave/components/variations` subdir, but it only
contains a build flag, no code, so I just added the the
pref registration to the main process function.

We've always discussed this as an opt-out. so default to `true`.