WIP: Add WeeklyEventStorage.
This is a helper class for recording events for later P3A reporting.

It remembers a sequence on enum values for a week, and can return
the most recently-added value within that time period, if any.

NB: This version fails to link with undefined symbol errors for
the templated method functions.