P3A: Make metric name list private.
Move the list of reported metric (histogram) names to a separate
compilation unit. This keeps them in a separate file for review
purposes but avoids the danger of including multiple copies.

Provide a function to register an observer callback as is needed
by BraveP3AService, since it can no longer iterate over the
constexpr HashSet directly.

Provide a membership function as well. Since this is no longer
an internal component of BraveP3AService, remove the indirection
through the BraveP3ALogStore delegate and call the component function
directly.